refactor(agent-skills): establish Skillpack projection boundary
