+++ lib/hive/version.rb
@@ -1,0 +1,3 @@
+module Hive
+  VERSION = "0.6.9".freeze
+end
