+++ test/unit/commands/doctor_managed_skills_test.rb
@@ -4,6 +4,7 @@ require "fileutils"
 require "json"
 require "stringio"
 require "hive/commands/doctor"
+require "hive/agent_skills/inspector"
 
 class HiveCommandsDoctorManagedSkillsTest < Minitest::Test
   include HiveTestHelper
